Change BigDecimal to Integer for avoid bug of class type not match in Seata 0.6.1

+++ cloud-seata-nacos/order-service/src/main/java/io/github/helloworlde/order/model/Order.java
@@ -5,8 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 /**
  * 订单
  *
@@ -26,5 +24,5 @@ public class Order {
 
     private OrderStatus status;
 
-    private BigDecimal payAmount;
+    private Integer payAmount;
 }
